Rename main script
