+++ ihp-sg13g2/libs.tech/ngspice/testing/devices/tests/test_devices.py
@@ -2,7 +2,7 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from models_verifier.mdm_verifier import MdmVerifier
+from models_verifier.models_verifier import MdmVerifier
 
 CASES = [
     ("nmos_lv", "mos/nmos_lv/sg13_lv_nmos.yaml"),
